requirements.txt
bitstring==3.1.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
dimod==0.8.9
dwave-cloud-client==0.5.3
dwave-hybrid==0.2.0
dwave-neal==0.4.5
dwave-networkx==0.7.1
dwave-ocean-sdk==1.3.0
dwave-qbsolv==0.2.10
dwave-system==0.7.2
dwave-tabu==0.1.3
dwavebinarycsp==0.0.10
homebase==1.0.1
idna==2.8
jsonschema==2.6.0
minorminer==0.1.8
networkx==2.2
numpy==1.15.4
ortools==6.10.6025
penaltymodel==0.15.4
penaltymodel-cache==0.3.3
penaltymodel-mip==0.1.4
pkg-resources==0.0.0
plucky==0.4.3
protobuf==3.7.1
PySocks==1.6.8
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
